pipelines/ETL/pyproject.toml
[tool.poetry]
name = "data-fetch"
version = "0.1.0"
description = ""
authors = ["Joywalker <alexandrurazvant@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-unsplash = "^1.2.5"
pexels-api = "^1.0.1"
structlog = "^24.4.0"
python-dotenv = "^1.0.1"
aiohttp = "^3.11.7"
motor = "^3.6.0"
pydantic = "^2.10.0"
tenacity = "^9.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
